pyproject.toml
[project]
name = "private_rag"
version = "0.1.0"
description = "Private RAG"
authors = [
    {name = "Belhadj Said",email = "sbelhadj@et.esiea.fr"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "llama-index (>=0.14.15,<0.15.0)",
    "unstructured[pdf] (>=0.21.5,<0.22.0)",
    "llama-index-readers-file (>=0.5.6,<0.6.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "chromadb (>=1.5.2,<2.0.0)",
    "llama-index-vector-stores-chroma (>=0.5.5,<0.6.0)",
    "llama-index-llms-litellm (>=0.6.3,<0.7.0)",
    "llama-index-embeddings-huggingface (>=0.6.1,<0.7.0)",
    "fastapi[standard] (>=0.134.0,<0.135.0)"
]
[tool.poetry]
packages = [{ include = "private_rag" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"